docs(memory): CONTRIBUTING.md Rule 2 contradicted the contract, the code and two green tests
Review raised the SPI javadoc and CONTRIBUTING.md disagreeing about double close, and suggested a
follow-up. Checking which one is right settled it in the opposite direction from what both of us
assumed, so it is fixed here rather than deferred.

AbstractLoanedBuffer.close() opens its CAS loop with `if (prev <= 0) { return; }`. A close on an
already-released buffer is a no-op. The SPI javadoc requires exactly that, and it is not merely
documented: AbstractLoanedBufferTest ("Double close is idempotent -- does not release twice") and
CommunityLoanedBufferTest ("double close() after single allocate() does not throw") have both been
green this whole time. CONTRIBUTING.md was contradicting the contract, the implementation and two
passing tests at once -- and my own previous commit message repeated its claim, which is how a wrong
rule stays alive.

Rule 2 now states the real constraint: BALANCE, NOT REPETITION. Every retain() needs its own close(),
because close() releases only at zero; the use-after-free comes from handing a buffer to a subtask
without retaining first, not from closing twice.

The exchange keeps its CAS guard, with the reason rewritten to the true one. It is not protecting
against a decrement that cannot happen; it is protecting against a MemoryAllocator implementation
this repository cannot read, since the SPI is what a driver implements. Guarding costs one boolean.

Also from this review round, and the second item is a good question that turned out to have a precise
answer. The two int additions in findHeaderStart are now long throughout -- the method returns a long
and MemorySegment indexes in longs, so computing in int and widening at the return was the wrong way
round even where nothing can overflow; doing the arithmetic in the type the result is used at removes
the question rather than answering it in a comment.

The "Upgrade" literal now goes through the existing constant, and the comment says why that is
correct rather than merely deduplicating: RFC 9110 §7.6.1 defines `Connection` as a list of HEADER
FIELD NAMES, so `Connection: Upgrade` literally names the `Upgrade` header the same code reads. They
are one token, not two that happen to match.

The SHA-1 citation is tightened to RFC 6455 §4.2.2 STEP 5.4. The bare section reference was
defensible but easy to read as wrong: steps 1-4 of §4.2.2 never mention SHA-1, so anyone checking the
section from the top hits /origin/ and /key/ and finds nothing. Verified against the RFC text rather
than from memory -- step 5.4 is explicit, "taking the SHA-1 hash of this concatenated value" after
concatenating /key/ with the GUID. The comment now names the step and says why the imprecise citation
was worth fixing.

Verified: full reactor mvn clean install green with NO skip flags; 3830 tests, 0 failures, 18
skipped; ExerisArchitectureTest 13/13; KernelTierBanArchitectureTest 5/5; the two LoanedBuffer tests
cited in the correction re-run and green; WebSocket suite 14/14; lint clean on the changed module.

+++ exeris-kernel-community/src/main/java/eu/exeris/kernel/community/websocket/CommunityWebSocketUpgradeRequest.java
@@ -99,10 +99,11 @@ final class CommunityWebSocketUpgradeRequest {
     }
 
     private static long findHeaderStart(MemorySegment segment, int length) {
-        // `i + 1L` rather than `i + 1`: the comparison is against an int length so the addition
-        // cannot actually overflow here, but writing the widening explicitly keeps the analyser and
-        // the next reader from having to re-derive that.
-        for (int i = 0; i + 1L < length; i++) {
+        // Long throughout: the method returns a long offset and MemorySegment indexes in longs, so
+        // computing in int and widening at the return was the wrong way round even where the values
+        // cannot overflow. Doing the arithmetic in the type the result is used at removes the
+        // question instead of answering it in a comment.
+        for (long i = 0; i + 1 < length; i++) {
             if (segment.get(ValueLayout.JAVA_BYTE, i) == '\r'
                     && segment.get(ValueLayout.JAVA_BYTE, i + 1) == '\n') {
                 return i + 2;
